requirements.txt
pandas
scikit-learn
python-telegram-bot